+++ __tests__/basic.js
@@ -13,7 +13,7 @@ test('setup', async () => {
   const { driver, pageObjects } = await setupWebDriver();
 
   await driver.wait(uiConnected(), timeouts.directLine);
-  await pageObjects.sendMessageViaSendBox('layout carousel');
+  await pageObjects.sendMessageViaSendBox('layout carousel', { waitForSend: true });
 
   await driver.wait(minNumActivitiesShown(2), 2000);
   await driver.wait(allImagesLoaded(), 2000);
@@ -27,7 +27,7 @@ test('long URLs with break-word', async () => {
   const { driver, pageObjects } = await setupWebDriver();
 
   await driver.wait(uiConnected(), timeouts.directLine);
-  await pageObjects.sendMessageViaSendBox('https://subdomain.domain.com/pathname0/pathname1/pathname2/pathname3/pathname4/')
+  await pageObjects.sendMessageViaSendBox('https://subdomain.domain.com/pathname0/pathname1/pathname2/pathname3/pathname4/', { waitForSend: true })
 
   await driver.wait(minNumActivitiesShown(2), 2000);
   await driver.wait(allImagesLoaded(), 2000);
@@ -43,7 +43,7 @@ test('long URLs with break-all', async () => {
   const { driver, pageObjects } = await setupWebDriver({ props: WEB_CHAT_PROPS });
 
   await driver.wait(uiConnected(), timeouts.directLine);
-  await pageObjects.sendMessageViaSendBox('https://subdomain.domain.com/pathname0/pathname1/pathname2/pathname3/pathname4/')
+  await pageObjects.sendMessageViaSendBox('https://subdomain.domain.com/pathname0/pathname1/pathname2/pathname3/pathname4/', { waitForSend: true })
 
   await driver.wait(minNumActivitiesShown(2), 2000);
   await driver.wait(allImagesLoaded(), 2000);
@@ -58,7 +58,7 @@ test('long URLs with keep-all', async () => {
 
   const { driver, pageObjects } = await setupWebDriver({ props: WEB_CHAT_PROPS });
 
-  await pageObjects.sendMessageViaSendBox('箸より重いものを持ったことがない箸より重いものを持ったことがない');
+  await pageObjects.sendMessageViaSendBox('箸より重いものを持ったことがない箸より重いものを持ったことがない', { waitForSend: true });
 
   await driver.wait(minNumActivitiesShown(2), 2000);
   await driver.wait(allImagesLoaded(), 2000);
@@ -72,7 +72,7 @@ test('unknown activities do not render anything in the transcript', async () => 
   const { driver, pageObjects } = await setupWebDriver();
 
   await driver.wait(uiConnected(), timeouts.directLine);
-  await pageObjects.sendMessageViaSendBox('unknown activity');
+  await pageObjects.sendMessageViaSendBox('unknown activity', { waitForSend: true });
 
   await driver.wait(minNumActivitiesShown(1), 2000);