+++ __tests__/offlineUI.js
@@ -1,6 +1,8 @@
 import { By, Condition, Key } from 'selenium-webdriver';
 
 import { imageSnapshotOptions, timeouts } from './constants.json';
+import actionDispatched from './setup/conditions/actionDispatched';
+import minNumActivitiesShown from './setup/conditions/minNumActivitiesShown';
 import staticSpinner from './setup/assets/staticSpinner';
 import uiConnected from './setup/conditions/uiConnected';
 
@@ -19,12 +21,9 @@ const allOutgoingMessagesFailed = new Condition('All outgoing messages to fail s
 });
 
 describe('offline UI', async () => {
-  test('should show "taking longer than usual to connect" UI when connection is slow', async () => {
-
-    const WEB_CHAT_PROPS = { spinnerAnimationBackgroundImage: staticSpinner };
+  test('should show "Taking longer than usual to connect" UI when connection is slow', async () => {
 
     const { driver } = await setupWebDriver({
-      props: { WEB_CHAT_PROPS },
       createDirectLine: options => {
         // This part of code is running in the JavaScript VM in Chromium.
         // This variable must be declared within scope
@@ -50,17 +49,25 @@ describe('offline UI', async () => {
         };
       },
       pingBotOnLoad: false,
-      setup: () => new Promise(resolve => {
-        const scriptElement = document.createElement('script');
-
-        scriptElement.onload = resolve;
-        scriptElement.setAttribute('src', 'https://unpkg.com/core-js@2.6.3/client/core.min.js');
-
-        document.head.appendChild(scriptElement);
+      setup: () =>
+      Promise.all([
+        window.WebChatTest.loadScript('https://unpkg.com/core-js@2.6.3/client/core.min.js'),
+        window.WebChatTest.loadScript('https://unpkg.com/lolex@4.0.1/lolex.js'),
+      ]).then(() => {
+        window.WebChatTest.clock = lolex.install();
       })
     });
 
-    await driver.sleep(15000);
+    await driver.executeScript(() => {
+      window.WebChatTest.clock.tick(400); // "Connecting" will be gone after 400ms, turning into "Taking longer than usual to connect"
+      window.WebChatTest.clock.tick(14600); // Go to t=15s
+    });
+
+    await driver.executeScript(() => {
+      window.WebChatTest.clock.tick(1); // Shortly after 15s, it will show "Taking longer than usual to connect"
+    });
+
+    await driver.wait(actionDispatched('DIRECT_LINE/CONNECT_STILL_PENDING'), timeouts.directLine);
 
     const base64PNG = await driver.takeScreenshot();
 
@@ -186,11 +193,10 @@ describe('offline UI', async () => {
   });
 
   test('should display the "Connecting..." connectivity status when connecting for the first time', async() => {
-    const WEB_CHAT_PROPS = { spinnerAnimationBackgroundImage: staticSpinner };
+    const WEB_CHAT_PROPS = { styleOptions: { spinnerAnimationBackgroundImage: staticSpinner } };
 
     const { driver } = await setupWebDriver({
-      props: WEB_CHAT_PROPS,
-      createDirectline: options => {
+      createDirectLine: options => {
         // This part of code is running in the JavaScript VM in Chromium.
         // This Direct Line Connection Status variable must be declared within scope
         const UNINITIALIZED = 0;
@@ -215,6 +221,7 @@ describe('offline UI', async () => {
         };
       },
       pingBotOnLoad: false,
+      props: WEB_CHAT_PROPS,
       setup: () => new Promise(resolve => {
         const scriptElement = document.createElement('script');
 
@@ -227,14 +234,14 @@ describe('offline UI', async () => {
     });
 
     const base64PNG = await driver.takeScreenshot();
+
     expect(base64PNG).toMatchImageSnapshot(imageSnapshotOptions);
   });
 
   test('should display "Network interruption occurred. Reconnecting…" status when connection is interrupted', async () => {
-    const WEB_CHAT_PROPS = { spinnerAnimationBackgroundImage: staticSpinner };
+    const WEB_CHAT_PROPS = { styleOptions: { spinnerAnimationBackgroundImage: staticSpinner } };
 
     const { driver } = await setupWebDriver({
-      props: { WEB_CHAT_PROPS },
       createDirectLine: options => {
         // This part of code is running in the JavaScript VM in Chromium.
         // These Direct Line Connection Status variables must be declared within scope
@@ -263,18 +270,33 @@ describe('offline UI', async () => {
         };
       },
       pingBotOnLoad: false,
-      setup: () => new Promise(resolve => {
-        const scriptElement = document.createElement('script');
-
-        scriptElement.onload = resolve;
-        scriptElement.setAttribute('src', 'https://unpkg.com/core-js@2.6.3/client/core.min.js');
-
-        document.head.appendChild(scriptElement);
+      props: WEB_CHAT_PROPS,
+      setup: () =>
+      Promise.all([
+        window.WebChatTest.loadScript('https://unpkg.com/core-js@2.6.3/client/core.min.js'),
+        window.WebChatTest.loadScript('https://unpkg.com/lolex@4.0.1/lolex.js')
+      ]).then(() => {
+        window.WebChatTest.clock = lolex.install();
       })
     });
 
-    await driver.sleep(600);
+    await driver.wait(actionDispatched('DIRECT_LINE/CONNECT_PENDING'), timeouts.directLine);
+    await driver.wait(actionDispatched('DIRECT_LINE/CONNECT_FULFILLED'), timeouts.directLine);
+    await driver.wait(actionDispatched('DIRECT_LINE/CONNECT_PENDING'), timeouts.directLine);
+
+    await driver.executeScript(() => {
+      window.WebChatTest.clock.tick(400); // "Connecting" will be gone after 400ms, turning into "Network interruption occured"
+      window.WebChatTest.clock.tick(200);
+    });
+
+    // TODO: [P4] Understand why we need to fire tick() using two cross-VM calls
+    //       When we put everything in a single cross-VM call, the last tick has no effect
+    await driver.executeScript(() => {
+      window.WebChatTest.clock.tick(1); // Shortly after 15s, it will show "Network interruption occured."
+    });
+
     const base64PNG = await driver.takeScreenshot();
+
     expect(base64PNG).toMatchImageSnapshot(imageSnapshotOptions);
   });
 
@@ -308,17 +330,51 @@ describe('offline UI', async () => {
         };
       },
       pingBotOnLoad: false,
-      setup: () => new Promise(resolve => {
-        const scriptElement = document.createElement('script');
-
-        scriptElement.onload = resolve;
-        scriptElement.setAttribute('src', 'https://unpkg.com/core-js@2.6.3/client/core.min.js');
-
-        document.head.appendChild(scriptElement);
-      })
+      setup: () =>
+        Promise.all([
+          window.WebChatTest.loadScript('https://unpkg.com/core-js@2.6.3/client/core.min.js'),
+          window.WebChatTest.loadScript('https://unpkg.com/lolex@4.0.1/lolex.js')
+        ]).then(() => {
+          window.WebChatTest.clock = lolex.install();
+        })
     });
 
-    await driver.sleep(17000);
+    await driver.wait(actionDispatched('DIRECT_LINE/RECONNECT_PENDING'), timeouts.directLine);
+
+    await driver.executeScript(() => {
+      window.WebChatTest.clock.tick(400); // "Connecting" will be gone after 400ms
+      window.WebChatTest.clock.tick(14600); // Go to t=15s
+    });
+
+    // TODO: [P4] Understand why we need to fire tick() using two cross-VM calls
+    //       When we put everything in a single cross-VM call, the last tick has no effect
+    await driver.executeScript(() => {
+      window.WebChatTest.clock.tick(1); // Shortly after 15s, it will show "Taking longer than usual to connect"
+    });
+
+    const base64PNG = await driver.takeScreenshot();
+
+    expect(base64PNG).toMatchImageSnapshot(imageSnapshotOptions);
+  });
+
+  test('should show "Render error" connectivity status when a JavaScript error is present in the code.', async () => {
+    const { driver, pageObjects } = await setupWebDriver({
+      storeMiddleware: ({ dispatch }) => next => action => {
+        if (action.type === 'DIRECT_LINE/INCOMING_ACTIVITY' && action.payload.activity && action.payload.activity.text === 'error') {
+          dispatch({
+            type: 'DIRECT_LINE/POST_ACTIVITY',
+            payload: {}
+          });
+        }
+
+        return next(action);
+      }
+    });
+
+    await driver.wait(uiConnected(), timeouts.directLine);
+    await pageObjects.sendMessageViaSendBox('error', { waitForSend: false });
+    await driver.wait(minNumActivitiesShown(2), timeouts.directLine);
+    await driver.wait(actionDispatched('WEB_CHAT/SAGA_ERROR'), timeouts.directLine);
 
     const base64PNG = await driver.takeScreenshot();